requirements.txt
langchain
streamlit
openai
tabulate
timm
transformers
python-dotenv
torch
Pillow